+++ app/src/main/java/physics/PhysicalPositional.java
@@ -5,41 +5,41 @@ import control.Game;
 
 public abstract class PhysicalPositional extends Positional {
 
-    private double resistance;
-    private double verticalAirResistance;
-    private double horizontalAirResistance;
-    private double frictionalResistance;
+    private float resistance;
+    private float verticalAirResistance;
+    private float horizontalAirResistance;
+    private float frictionalResistance;
     int verticalCollision;
     int horizontalCollision;
     protected Hitbox[] hitbox =  { new Hitbox(0, 0, 0, 0) };
-    protected double velocityX = 0;
-    protected double velocityY = 0;
-    protected double mass = 1;
+    protected float velocityX = 0;
+    protected float velocityY = 0;
+    protected float mass = 1;
 
-    public void setResistance(double resistance, Game game) {
+    public void setResistance(float resistance, Game game) {
         if (!(game.getPhysics() instanceof RoamerPhysics)) throw (new IllegalArgumentException());
         if (resistance < 0 || resistance > 1) throw (new IllegalArgumentException());
         this.resistance = resistance;
     }
 
-    public void setVerticalAirResistance(double resistance, Game game) {
+    public void setVerticalAirResistance(float resistance, Game game) {
         if (!(game.getPhysics() instanceof PlatformerPhysics)) throw (new IllegalArgumentException());
         if (resistance < 0 || resistance > 1) throw (new IllegalArgumentException());
         this.verticalAirResistance = resistance;
     }
 
-    public void setHorizontalAirResistance(double resistance, Game game) {
+    public void setHorizontalAirResistance(float resistance, Game game) {
         if (!(game.getPhysics() instanceof PlatformerPhysics)) throw (new IllegalArgumentException());
         if (resistance < 0 || resistance > 1) throw (new IllegalArgumentException());
         this.horizontalAirResistance = resistance;
     }
 
-    public void setFrictionalResistance(double resistance, Game game) {
+    public void setFrictionalResistance(float resistance, Game game) {
         if (resistance < 0 || resistance > 1) throw (new IllegalArgumentException());
         this.frictionalResistance = resistance;
     }
 
-    double getSpecVerticalResistance(Physics physics) {
+    float getSpecVerticalResistance(Physics physics) {
         if (physics instanceof PlatformerPhysics) {
             return verticalAirResistance;
         } else if (physics instanceof RoamerPhysics) {
@@ -51,7 +51,7 @@ public abstract class PhysicalPositional extends Positional {
         throw (new IllegalArgumentException());
     }
 
-    double getSpecHorizontalResistance(Physics physics) {
+    float getSpecHorizontalResistance(Physics physics) {
         if (physics instanceof PlatformerPhysics) {
             if (verticalCollision != 0)
                 return horizontalAirResistance + frictionalResistance;
@@ -74,27 +74,27 @@ public abstract class PhysicalPositional extends Positional {
         this.hitbox = hitbox;
     }
 
-    public double getVelocityX() {
+    public float getVelocityX() {
         return velocityX;
     }
 
-    public void setVelocityX(double velocityX) {
+    public void setVelocityX(float velocityX) {
         this.velocityX = velocityX;
     }
 
-    public void incVelocityX(double inc) {
+    public void incVelocityX(float inc) {
         this.velocityX += inc;
     }
 
-    public double getVelocityY() {
+    public float getVelocityY() {
         return velocityY;
     }
 
-    public void setVelocityY(double velocityY) {
+    public void setVelocityY(float velocityY) {
         this.velocityY = velocityY;
     }
 
-    public void incVelocityY(double inc) {
+    public void incVelocityY(float inc) {
         this.velocityY += inc;
     }
 
@@ -103,7 +103,7 @@ public abstract class PhysicalPositional extends Positional {
         y += velocityY;
     }
 
-    void changeVelocityWithResistance(double resistanceX, double resistanceY) {
+    void changeVelocityWithResistance(float resistanceX, float resistanceY) {
         velocityX *= 1 - resistanceX;
         velocityY *= 1 - resistanceY;
     }
@@ -116,11 +116,11 @@ public abstract class PhysicalPositional extends Positional {
         return horizontalCollision;
     }
 
-    public double getMass() {
+    public float getMass() {
         return mass;
     }
 
-    public void setMass(double mass) {
+    public void setMass(float mass) {
         this.mass = mass;
     }
 }
